Final Day Update 1
Guard Navigation: The issue I cannot seem to solve is that when the box
collider on the guard touches something, it breaks the navmesh agent.
The current solution was to reduce the size of the guard and some
objects colliders to allow room.

Guard Speed: The guards speed will increase from 7 to 12 when the player
is in his sights. It will reduce back to 7 when player is out of sight.

HUD: The current hud is very simple and has a placeholder image. The how
to play appears for 2.5 seconds before disappearing, the objective
changes based on the trigger of the glass case and countClick.

Glass Animation: Fixed, should work without fail now.