Update
The Guard Dog is a bit smoother, but also has a larger radius. The issue
where he flips out has to do with the Rigidbody and it is not perfect.

The dog guard is inactive to test the text on the glass.

There are also a cat paw and a mini guard with a texture on him, they
were both also there for testing purposes.